Docker/requirements.txt
commlib-py==0.11.1
redis
hiredis
paho-mqtt
pika